rename ability to skill
